feat: remove unused class and student pages; streamline admin classes structure and improve class management functionality
