--- app/admin/classes/create/page.tsx
@@ -1,220 +1,0 @@
-"use client";
-
-import type React from "react";
-
-import { useState } from "react";
-import { useRouter } from "next/navigation";
-import Link from "next/link";
-import { ArrowLeft, Save } from "lucide-react";
-
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Textarea } from "@/components/ui/textarea";
-import { toast } from "sonner";
-
-// Mock departments
-const departments = [
-  "Computer Science",
-  "Mathematics",
-  "Physics",
-  "Chemistry",
-  "Biology",
-  "English",
-  "History",
-  "Geography",
-];
-
-// Mock teachers
-const teachers = [
-  { id: "1", name: "John Smith" },
-  { id: "2", name: "Sarah Johnson" },
-  { id: "3", name: "Michael Brown" },
-  { id: "4", name: "Emily Davis" },
-  { id: "5", name: "David Wilson" },
-];
-
-export default function CreateClassPage() {
-  const router = useRouter();
-
-  const [formData, setFormData] = useState({
-    name: "",
-    department: "",
-    teacherId: "",
-    description: "",
-    schedule: "",
-    room: "",
-  });
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({ ...prev, [name]: value }));
-  };
-
-  const handleDepartmentChange = (value: string) => {
-    setFormData((prev) => ({ ...prev, department: value }));
-  };
-
-  const handleTeacherChange = (value: string) => {
-    setFormData((prev) => ({ ...prev, teacherId: value }));
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsSubmitting(true);
-
-    // Simulate API call
-    setTimeout(() => {
-      toast.success("Class created successfully");
-      setIsSubmitting(false);
-      router.push("/admin/classes");
-    }, 1000);
-  };
-
-  return (
-    <div className="space-y-6">
-      <div className="flex items-center gap-2">
-        <Button variant="ghost" size="icon" asChild>
-          <Link href="/admin/classes">
-            <ArrowLeft className="h-4 w-4" />
-            <span className="sr-only">Back</span>
-          </Link>
-        </Button>
-        <div>
-          <h1 className="text-3xl font-bold tracking-tight dark:text-white">
-            Add New Class
-          </h1>
-          <p className="text-muted-foreground dark:text-gray-400">
-            Create a new class
-          </p>
-        </div>
-      </div>
-
-      <form onSubmit={handleSubmit}>
-        <Card>
-          <CardHeader>
-            <CardTitle>Class Information</CardTitle>
-            <CardDescription>Enter the class details</CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-6">
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-              <div className="space-y-2">
-                <Label htmlFor="name">Class Name</Label>
-                <Input
-                  id="name"
-                  name="name"
-                  value={formData.name}
-                  onChange={handleChange}
-                  required
-                  placeholder="e.g. Introduction to Programming"
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="department">Department</Label>
-                <Select
-                  value={formData.department}
-                  onValueChange={handleDepartmentChange}
-                >
-                  <SelectTrigger id="department">
-                    <SelectValue placeholder="Select department" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {departments.map((dept) => (
-                      <SelectItem key={dept} value={dept}>
-                        {dept}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="teacher">Teacher</Label>
-                <Select
-                  value={formData.teacherId}
-                  onValueChange={handleTeacherChange}
-                >
-                  <SelectTrigger id="teacher">
-                    <SelectValue placeholder="Select teacher" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {teachers.map((teacher) => (
-                      <SelectItem key={teacher.id} value={teacher.id}>
-                        {teacher.name}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="room">Room</Label>
-                <Input
-                  id="room"
-                  name="room"
-                  value={formData.room}
-                  onChange={handleChange}
-                  required
-                  placeholder="e.g. CS-101"
-                />
-              </div>
-              <div className="space-y-2 md:col-span-2">
-                <Label htmlFor="schedule">Schedule</Label>
-                <Input
-                  id="schedule"
-                  name="schedule"
-                  value={formData.schedule}
-                  onChange={handleChange}
-                  required
-                  placeholder="e.g. Monday, Wednesday, Friday 10:00 AM - 11:30 AM"
-                />
-              </div>
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="description">Description</Label>
-              <Textarea
-                id="description"
-                name="description"
-                value={formData.description}
-                onChange={handleChange}
-                rows={4}
-                placeholder="Enter class description..."
-              />
-            </div>
-            <div className="flex justify-end gap-3">
-              <Button variant="outline" asChild>
-                <Link href="/admin/classes">Cancel</Link>
-              </Button>
-              <Button type="submit" disabled={isSubmitting}>
-                {isSubmitting ? (
-                  "Creating..."
-                ) : (
-                  <>
-                    <Save className="h-4 w-4 mr-2" />
-                    Create Class
-                  </>
-                )}
-              </Button>
-            </div>
-          </CardContent>
-        </Card>
-      </form>
-    </div>
-  );
-}
