--- app/admin/classes/layout.tsx
@@ -1,9 +1,0 @@
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-    // const classes = await getAllClassesWithStudentCount();
-
-  return <div>{children}</div>;
-}
